python/tools/check_cid_timeout/requirement.txt
beautifulsoup4==4.6.3
bs4==0.0.1
gevent==1.3.6
greenlet==0.4.15
html5lib==1.0.1
numpy==1.15.2
pandas==0.23.4
prettytable==0.7.2
PyYAML==3.13
requests==2.19.1
urllib3==1.23